HDDS-14021. RDBStore should not flush on rocksdb close

+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/utils/db/RDBStore.java
@@ -76,6 +76,7 @@ public class RDBStore implements DBStore {
   private final long maxDbUpdatesSizeThreshold;
   private final ManagedDBOptions dbOptions;
   private final ManagedStatistics statistics;
+  private final boolean readOnly;
 
   @SuppressWarnings("parameternumber")
   RDBStore(File dbFile, ManagedDBOptions dbOptions, ManagedStatistics statistics,
@@ -113,6 +114,7 @@ public class RDBStore implements DBStore {
 
       db = RocksDatabase.open(dbFile, dbOptions, writeOptions,
           families, readOnly);
+      this.readOnly = readOnly;
 
       // dbOptions.statistics() only contribute to part of RocksDB metrics in
       // Ozone. Enable RocksDB metrics even dbOptions.statistics() is off.
@@ -253,7 +255,9 @@ public class RDBStore implements DBStore {
     }
     try {
       // Flush to ensure all data is persisted to disk before closing.
-      flushDB();
+      if (!readOnly) {
+        flushDB();
+      }
       LOG.debug("Successfully flushed DB before close");
     } catch (Exception e) {
       LOG.warn("Failed to flush DB before close", e);
@@ -359,14 +363,7 @@ public class RDBStore implements DBStore {
    */
   @Override
   public void dropTable(String tableName) throws RocksDatabaseException {
-    ColumnFamily columnFamily = db.getColumnFamily(tableName);
-    if (columnFamily != null) {
-      try {
-        db.getManagedRocksDb().get().dropColumnFamily(columnFamily.getHandle());
-      } catch (RocksDBException e) {
-        throw new RocksDatabaseException("Failed to drop " + tableName, e);
-      }  
-    }
+    db.dropColumnFamily(tableName);
   }
 
   public Collection<ColumnFamily> getColumnFamilies() {
